Add Support for Calculating Block Level Depth Corrected Pressures
This commit calculates the block-level depth-corrected phase
pressures BPPO, BPPG, and BPPW.  We use the simulation
configuration's DatumDepth feature to extract the reference depths
for the FIPNUM region set per cell and accumulate average phase
densities per PVT regions using the RegionPhasePoreVolAverage
utility.  To that end, add a new optional<> data member of this type
to the generic output module and initialise this if the input model
requests such summary output.