Improve parallel network test.
The PID part of the timestep controller triggers and results in timestep
sizes depending on the state of the simulation. This can lead to different
timestepping for parallel and serial runs. Using only newton iteration count
for timestep calculations addresses this.

Also, tighten CNV tolerances to avoid solution differences that are accepted by
the simulator as within tolerances, but outside comparison tolerances.