fixed: do not dereference invalid iterator
if a well has connections added in an action, the well perf data
is not yet initialized in the call to updateEclWells. it will however
be reinitialized before the next time step so simply adding a bounds
check here avoid the invalid dereference and the connection factor
will take effect through the reinitialization.