[bugfix] Always broadcast NNCs from input.
This was only done if there were NNCs from aquifers, but it should
always be done whenever there are NNCs.

Those NNCs also hold information of (additional) transmissibilities
for a grid intersection. While the intersection is already represented
by the grid, the transmissibility is not. This PR fixes this and makes
the NNcs available on all processes.