Update group target reduction during local solves
Since the globalGRUP vector can not be communicated during local solves,
it may introduce differences in non-linear behaviour between MPI solves.
It should however converge to the same solution.

Without this PR the swithing to and from GRUP during local solves may give wrong results
as the group target reduction is not updated.