changed: make FlowProblem(Blackoil|Comp) return a reference to EclWriter
the pointer is allocated in the ctor, so the fact that it is a unique_ptr
can be kept as an implementation detail. furthermore, the previous
accessor did not properly enforce constness, as a const unique_ptr<EclWriter>
only applies const to the pointer class, not the data the pointer points to.