Remove unneeded broadcast() overload.
Note that this changes the order of arguments for some calls: now the
root always comes first, there is no possible mixup between the root
argument and the data arguments.

Most changes are due to no longer having a default argument
for the root. I consider that a good thing in itself.