Add End-of-Episode Predicate to Flow Problem Classes
This is arguably a hack.  The default implementation in terms of
Simulator<>::episodeWillBeOver() is known to have non-trivial false
negatives in certain regression tests.  These are in turn possibly
related to using the simulation start time point (often a large
number of seconds) and subtracting values of similar size.  The
custom implementation in FlowProblemBlackoil<> uses a floating-point
comparison involving Schedule::seconds() instead, thereby reducing
the size of the numbers involved.  We're still vulnerable to
"catastrophic cancellation" errors, but potentially less so.

A complete fix for the problem may involve switching to representing
the simulated time variable as an integer instead of a
floating-point type.