Generalize initial value retrieval of reduction operators
The initial value retrieval in the reduction operator is generalized to allow the return from a template argument. The reason behind this is that binary operators in the standard like deprecated (C++17) and removed (C++20) the return type from its traits, meaning that the return type cannot be inferred by purely looking at them.