Revert previous fix of MULTREGT for PINCH NNCs from OPTION(4)==ALL
That fix only fixed the situation where the multiplier was 0.

For other values we need to apply the multipliers to the
transmissibilities of the pinched out cells and the neighbor cells
before calculating the harmonic average. That is done in
processEclipseFormat in opm-grid and stored in pinchNNCs in the
EclipseState.