Chase Summary::Eval() API Change
In particular, switch to passing a collection of pointers through
the Summary::DynamicSimulatorState object instead of passing
references to individual state objects.  This is in preparation of
adding more state objects, in particular for summary quantities at
the region level without having to pay a large API cost for each of
those.

To this end also switch the output module's initialInPlace()
function to returning a pointer--with nullptr being "no such
value"--instead of a reference to an optional<>.